Move grounding tool to reusable parent tools dir
